First cut at Llama example
Notably, this assumes a new syntax for running static expressions using `:=`.  These are used to precompute the RoPE frequencies.

There are other approaches that could be taken, including reading these in from external weights and passing around as |freqs| arguments.

But in principle, the approach outlined here could/should work, where some expressions can be statically evaluated.  This is consistent with the existing overlapping static and dynamic namespaces, and with the static evaluation of arguments to static params of functions, such as `E*4`.

Supporting what's used in this example though would push static computation much further, requiring support for at least:
* `**`
* `Range`
* `/`
* `Outer` (outer product)
* `Cos`/`Sin`

Which in turn implies the ability to call functions statically.  Many questions still whether this is a good idea.